Forward walking of instructions in basic block
Forward-walking is completed to get information about the
state-structure access. Need to now compact it into only the direct
information I want, by walking the instruction list in reverse, and then
i'll flatten to see if there are any conflicts. A conflict would denote
that the access type is alive.